% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagram.R
\name{add_flow}
\alias{add_flow}
\title{Add a flow to a diagram}
\usage{
add_flow(diagram, name, from = NULL, to = NULL, rate)
}
\arguments{
\item{diagram}{StockFlowDiagram object}

\item{name}{Character name for the flow}

\item{from}{Character name of source stock (or NULL for inflow)}

\item{to}{Character name of target stock (or NULL for outflow)}

\item{rate}{Function computing flow rate, with signature
function(inputs, params). Can also be a formula (converted to function).}
}
\value{
Modified StockFlowDiagram object
}
\description{
Add a flow to a diagram
}
\examples{
\dontrun{
diagram <- stock_flow_diagram() \%>\%
  add_stock("S", initial = 1000) \%>\%
  add_stock("I", initial = 1) \%>\%
  add_flow("infection",
    from = "S",
    to = "I",
    rate = function(inputs, params) {
      params$beta * inputs$S * inputs$I
    }
  )
}

}
