% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagram.R
\name{add_link}
\alias{add_link}
\title{Add a link from a variable to a flow}
\usage{
add_link(diagram, from, to)
}
\arguments{
\item{diagram}{StockFlowDiagram object}

\item{from}{Character name of source (stock or variable)}

\item{to}{Character name of target flow}
}
\value{
Modified StockFlowDiagram object
}
\description{
Links indicate which quantities (stocks or variables) a flow's
rate function depends on.
}
\examples{
\dontrun{
diagram <- diagram \%>\%
  add_link(from = "N", to = "infection")
}

}
