% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagram.R
\name{add_link_internal}
\alias{add_link_internal}
\title{Add a link to a diagram (internal function)}
\usage{
add_link_internal(diagram, name, from_id, from_type, to_flow)
}
\arguments{
\item{diagram}{StockFlowDiagram object}

\item{name}{Character name for the link}

\item{from_id}{Integer source ID}

\item{from_type}{Character source type ("stock", "variable", "sumvar")}

\item{to_flow}{Integer target flow ID}
}
\value{
Modified StockFlowDiagram object
}
\description{
Add a link to a diagram (internal function)
}
\keyword{internal}
