% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagram.R
\name{add_stock}
\alias{add_stock}
\title{Add a stock to a diagram}
\usage{
add_stock(diagram = NULL, name, initial = 0)
}
\arguments{
\item{diagram}{StockFlowDiagram object (or NULL to create new)}

\item{name}{Character name for the stock}

\item{initial}{Numeric initial value (default 0)}
}
\value{
Modified StockFlowDiagram object
}
\description{
Add a stock to a diagram
}
\examples{
\dontrun{
diagram <- stock_flow_diagram() \%>\%
  add_stock("S", initial = 1000) \%>\%
  add_stock("I", initial = 1)
}

}
