% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagram.R
\name{add_sum_variable}
\alias{add_sum_variable}
\title{Add a sum variable to a diagram}
\usage{
add_sum_variable(diagram, name, stocks)
}
\arguments{
\item{diagram}{StockFlowDiagram object}

\item{name}{Character name for the sum variable}

\item{stocks}{Character vector of stock names to sum}
}
\value{
Modified StockFlowDiagram object
}
\description{
Sum variables compute the sum of multiple stocks. Common use case
is computing total population N = S + E + I + R.
}
\examples{
\dontrun{
diagram <- diagram \%>\%
  add_sum_variable("N", stocks = c("S", "E", "I", "R"))
}

}
