% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagram.R
\name{add_variable}
\alias{add_variable}
\title{Add an auxiliary variable to a diagram}
\usage{
add_variable(diagram, name, expression)
}
\arguments{
\item{diagram}{StockFlowDiagram object}

\item{name}{Character name for the variable}

\item{expression}{Function computing the variable value, with signature
function(state, params, var_values)}
}
\value{
Modified StockFlowDiagram object
}
\description{
Auxiliary variables are computed values that can be used in flow
rate functions. They appear as circles in stock-flow diagrams.
}
\examples{
\dontrun{
diagram <- diagram \%>\%
  add_variable("force_of_infection",
    expression = function(state, params, var_values) {
      params$beta * state["I"] / sum(state)
    }
  )
}

}
