% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{check_interface_compatibility}
\alias{check_interface_compatibility}
\title{Check if two diagrams have compatible interfaces}
\usage{
check_interface_compatibility(left, right)
}
\arguments{
\item{left}{StockFlowDiagram or DecoratedCospan}

\item{right}{StockFlowDiagram or DecoratedCospan}
}
\value{
Logical indicating compatibility
}
\description{
Check if two diagrams have compatible interfaces
}
\keyword{internal}
