% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition.R
\name{close_diagram}
\alias{close_diagram}
\title{Close an OpenStockFlowDiagram}
\usage{
close_diagram(open_diagram)
}
\arguments{
\item{open_diagram}{OpenStockFlowDiagram object}
}
\value{
StockFlowDiagram object (the apex)
}
\description{
Forgets the interface information and returns the underlying StockFlowDiagram.
}
\examples{
\dontrun{
open_sir <- as_open_diagram(sir, right_interface = "I")
closed_sir <- close_diagram(open_sir)  # Same as original sir
}

}
