% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algebraic-api.R
\name{compile}
\alias{compile}
\title{Compile model builder into diagram}
\usage{
compile(builder, reset_ids = TRUE)
}
\arguments{
\item{builder}{model_builder object}

\item{reset_ids}{Logical; if TRUE, reset ID counter}
}
\value{
StockFlowDiagram object
}
\description{
Compile model builder into diagram
}
