% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{DiagramElement-class}
\alias{DiagramElement-class}
\title{Base class for all diagram elements}
\description{
Base class for all diagram elements
}
\section{Slots}{

\describe{
\item{\code{id}}{Integer identifier (unique within a diagram)}

\item{\code{name}}{Character name (human-readable label)}
}}

\keyword{internal}
