% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagram.R
\docType{data}
\name{.id_env}
\alias{.id_env}
\title{ID generator environment}
\format{
An object of class \code{environment} of length 1.
}
\usage{
.id_env
}
\description{
ID generator environment
}
\keyword{internal}
