% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition.R
\name{extract_interface}
\alias{extract_interface}
\title{Extract interface from diagram}
\usage{
extract_interface(diagram)
}
\arguments{
\item{diagram}{StockFlowDiagram or OpenStockFlowDiagram}
}
\value{
Integer vector of stock IDs
}
\description{
Gets the interface stock IDs from a diagram.
}
\keyword{internal}
