% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algebraic-api.R
\name{flow}
\alias{flow}
\title{Create a flow specification}
\usage{
flow(name, rate, from = NULL, to = NULL)
}
\arguments{
\item{name}{Character name for the flow}

\item{rate}{Function computing flow rate with signature function(inputs, params)}

\item{from}{Character name of source stock (set later with operator)}

\item{to}{Character name of target stock (set later with operator)}
}
\value{
A list with class "flow_spec"
}
\description{
Creates a flow specification that can be used with the algebraic API.
This does not add the flow to a diagram yet.
}
\examples{
\dontrun{
infection <- flow("infection",
  rate = function(inputs, params) {
    params$beta * inputs$S * inputs$I / inputs$N
  }
)
}

}
