% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagram.R
\name{flow_names}
\alias{flow_names}
\title{Get flow names}
\usage{
flow_names(diagram)
}
\arguments{
\item{diagram}{StockFlowDiagram object}
}
\value{
Character vector of flow names
}
\description{
Get flow names
}
