% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantics.R
\name{generate_ode}
\alias{generate_ode}
\title{Generate ODE function from stock-flow diagram}
\usage{
generate_ode(diagram)
}
\arguments{
\item{diagram}{StockFlowDiagram object}
}
\value{
Function with signature function(time, state, params)
  that returns list(derivatives)
}
\description{
Creates a function compatible with deSolve::ode() that computes
derivatives for all stocks.
}
\details{
Mathematical formula for stock σ:
dσ/dt = Σ_{f ∈ F(d)^{-1}(σ)} φ_f - Σ_{f ∈ F(u)^{-1}(σ)} φ_f

That is: rate of change = inflows - outflows
}
\examples{
\dontrun{
ode_func <- generate_ode(sir_diagram)
result <- deSolve::ode(
  y = initial_state(sir_diagram),
  times = seq(0, 100, by = 0.1),
  func = ode_func,
  parms = c(beta = 0.5, gamma = 0.1)
)
}

}
