% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{get_flow}
\alias{get_flow}
\title{Get flow by ID}
\usage{
get_flow(diagram, flow_id)
}
\arguments{
\item{diagram}{StockFlowDiagram object}

\item{flow_id}{Integer flow ID}
}
\value{
Flow object or NULL if not found
}
\description{
Get flow by ID
}
\keyword{internal}
