% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{get_flow_by_name}
\alias{get_flow_by_name}
\title{Get flow by name}
\usage{
get_flow_by_name(diagram, name)
}
\arguments{
\item{diagram}{StockFlowDiagram object}

\item{name}{Character flow name}
}
\value{
Flow object or NULL if not found
}
\description{
Get flow by name
}
\keyword{internal}
