% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{get_flows_from_stock}
\alias{get_flows_from_stock}
\title{Get flows with given source stock}
\usage{
get_flows_from_stock(diagram, stock_id)
}
\arguments{
\item{diagram}{StockFlowDiagram object}

\item{stock_id}{Integer stock ID}
}
\value{
Integer vector of flow IDs
}
\description{
Get flows with given source stock
}
\keyword{internal}
