% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{get_inflows}
\alias{get_inflows}
\title{Get inflows (flows with NA source)}
\usage{
get_inflows(diagram)
}
\arguments{
\item{diagram}{StockFlowDiagram object}
}
\value{
Integer vector of flow IDs
}
\description{
Get inflows (flows with NA source)
}
\keyword{internal}
