% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{get_link}
\alias{get_link}
\title{Get link by ID}
\usage{
get_link(diagram, link_id)
}
\arguments{
\item{diagram}{StockFlowDiagram object}

\item{link_id}{Integer link ID}
}
\value{
Link object or NULL if not found
}
\description{
Get link by ID
}
\keyword{internal}
