% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{get_links_from_source}
\alias{get_links_from_source}
\title{Get links from a source (stock or variable)}
\usage{
get_links_from_source(diagram, source_id)
}
\arguments{
\item{diagram}{StockFlowDiagram object}

\item{source_id}{Integer source ID}
}
\value{
Integer vector of link IDs
}
\description{
Get links from a source (stock or variable)
}
\keyword{internal}
