% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{get_links_to_flow}
\alias{get_links_to_flow}
\title{Get links targeting a flow}
\usage{
get_links_to_flow(diagram, flow_id)
}
\arguments{
\item{diagram}{StockFlowDiagram object}

\item{flow_id}{Integer flow ID}
}
\value{
Integer vector of link IDs
}
\description{
Get links targeting a flow
}
\keyword{internal}
