% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{get_outflows}
\alias{get_outflows}
\title{Get outflows (flows with NA target)}
\usage{
get_outflows(diagram)
}
\arguments{
\item{diagram}{StockFlowDiagram object}
}
\value{
Integer vector of flow IDs
}
\description{
Get outflows (flows with NA target)
}
\keyword{internal}
