% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{get_stock}
\alias{get_stock}
\title{Get stock by ID}
\usage{
get_stock(diagram, stock_id)
}
\arguments{
\item{diagram}{StockFlowDiagram object}

\item{stock_id}{Integer stock ID}
}
\value{
Stock object or NULL if not found
}
\description{
Get stock by ID
}
\keyword{internal}
