% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{get_stock_by_name}
\alias{get_stock_by_name}
\title{Get stock by name}
\usage{
get_stock_by_name(diagram, name)
}
\arguments{
\item{diagram}{StockFlowDiagram object}

\item{name}{Character stock name}
}
\value{
Stock object or NULL if not found
}
\description{
Get stock by name
}
\keyword{internal}
