% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{get_variable_by_name}
\alias{get_variable_by_name}
\title{Get variable by name}
\usage{
get_variable_by_name(diagram, name)
}
\arguments{
\item{diagram}{StockFlowDiagram object}

\item{name}{Character variable name}
}
\value{
AuxiliaryVariable or SumVariable object, or NULL if not found
}
\description{
Get variable by name
}
\keyword{internal}
