% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algebraic-api.R
\name{\%->\%}
\alias{\%->\%}
\title{Connect stocks via flows using algebraic notation}
\usage{
lhs \%->\% rhs
}
\arguments{
\item{lhs}{Left-hand side: stock_spec or flow_spec}

\item{rhs}{Right-hand side: flow_spec or stock_spec}
}
\value{
A partial flow connection (class "flow_connection")
}
\description{
The %->% operator creates a flow connection between two stocks.
It can be chained: stock1 %->% flow1 %->% stock2
}
\examples{
\dontrun{
# Define stocks
S <- stock("S", initial = 1000)
I <- stock("I", initial = 1)

# Define flow
infection <- flow("infection",
  rate = function(inputs, params) params$beta * inputs$S * inputs$I
)

# Connect: S \%->\% infection \%->\% I
connection <- S \%->\% infection \%->\% I
}

}
