% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagram.R
\name{initial_state}
\alias{initial_state}
\title{Get initial state vector}
\usage{
initial_state(diagram)
}
\arguments{
\item{diagram}{StockFlowDiagram object}
}
\value{
Named numeric vector of initial values
}
\description{
Get initial state vector
}
