% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algebraic-api.R
\name{model}
\alias{model}
\title{Start a diagram specification}
\usage{
model()
}
\value{
A diagram builder object
}
\description{
Alternative API that collects specifications and builds the diagram
at the end.
}
\examples{
\dontrun{
sir <- model() +
  stock("S", initial = 1000) +
  stock("I", initial = 1) +
  stock("R", initial = 0) +
  (S \%->\% flow("infection", rate = ...) \%->\% I) +
  (I \%->\% flow("recovery", rate = ...) \%->\% R)
}

}
