% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagram.R
\name{n_flows}
\alias{n_flows}
\title{Get number of flows in diagram}
\usage{
n_flows(diagram)
}
\arguments{
\item{diagram}{StockFlowDiagram object}
}
\value{
Integer count
}
\description{
Get number of flows in diagram
}
