% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagram.R
\name{n_links}
\alias{n_links}
\title{Get number of links in diagram}
\usage{
n_links(diagram)
}
\arguments{
\item{diagram}{StockFlowDiagram object}
}
\value{
Integer count
}
\description{
Get number of links in diagram
}
