% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagram.R
\name{n_stocks}
\alias{n_stocks}
\title{Get number of stocks in diagram}
\usage{
n_stocks(diagram)
}
\arguments{
\item{diagram}{StockFlowDiagram object}
}
\value{
Integer count
}
\description{
Get number of stocks in diagram
}
