% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{OpenStockFlowDiagram-class}
\alias{OpenStockFlowDiagram-class}
\title{Open Stock-Flow Diagram}
\description{
A decorated cospan where the legs are injective (one-to-one).
This ensures the diagram can be composed with others via pushout.
}
\section{Slots}{

\describe{
\item{\code{apex}}{StockFlowDiagram at the apex}

\item{\code{left_leg}}{Integer vector (injective) mapping left foot to apex}

\item{\code{right_leg}}{Integer vector (injective) mapping right foot to apex}

\item{\code{left_foot}}{Integer vector of interface stock IDs}

\item{\code{right_foot}}{Integer vector of interface stock IDs}
}}

