% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{phase_plot}
\alias{phase_plot}
\title{Create a phase plot of two stocks}
\usage{
phase_plot(solution, stock1, stock2, ...)
}
\arguments{
\item{solution}{stockflow_solution object}

\item{stock1}{Character name of first stock (x-axis)}

\item{stock2}{Character name of second stock (y-axis)}

\item{...}{Additional arguments}
}
\value{
ggplot object or base plot
}
\description{
Plots the trajectory in state space (stock1 vs stock2).
}
\examples{
\dontrun{
result <- solve_diagram(sir, ...)
phase_plot(result, "S", "I")
}

}
