% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_solution_base}
\alias{plot_solution_base}
\title{Plot solution using base graphics}
\usage{
plot_solution_base(solution, stocks, ...)
}
\description{
Plot solution using base graphics
}
\keyword{internal}
