% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_solution_ggplot}
\alias{plot_solution_ggplot}
\title{Plot solution using ggplot2}
\usage{
plot_solution_ggplot(solution, stocks, style = "line", ...)
}
\description{
Plot solution using ggplot2
}
\keyword{internal}
