% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algebraic-api.R
\name{+.model_builder}
\alias{+.model_builder}
\title{Add specification to model builder}
\usage{
\method{+}{model_builder}(builder, spec)
}
\arguments{
\item{builder}{model_builder object}

\item{spec}{Specification to add}
}
\value{
Updated model_builder
}
\description{
Add specification to model builder
}
