% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagram.R
\name{set_interface}
\alias{set_interface}
\title{Set the interface for composition}
\usage{
set_interface(diagram, stocks)
}
\arguments{
\item{diagram}{StockFlowDiagram object}

\item{stocks}{Character vector of stock names to expose}
}
\value{
Modified StockFlowDiagram object
}
\description{
The interface specifies which stocks are exposed for composition
with other diagrams.
}
\examples{
\dontrun{
diagram <- diagram \%>\%
  set_interface(c("S", "V"))
}

}
