% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagram.R
\name{stock_flow_diagram}
\alias{stock_flow_diagram}
\title{Create an empty stock-flow diagram}
\usage{
stock_flow_diagram(reset_ids = FALSE)
}
\arguments{
\item{reset_ids}{Logical; if TRUE, reset the ID counter (default FALSE)}
}
\value{
Empty StockFlowDiagram object
}
\description{
Create an empty stock-flow diagram
}
\examples{
\dontrun{
diagram <- stock_flow_diagram()
}

}
