% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagram.R
\name{stock_names}
\alias{stock_names}
\title{Get stock names}
\usage{
stock_names(diagram)
}
\arguments{
\item{diagram}{StockFlowDiagram object}
}
\value{
Character vector of stock names
}
\description{
Get stock names
}
