% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{summary_plot}
\alias{summary_plot}
\title{Summary visualization of solution}
\usage{
summary_plot(solution, ...)
}
\arguments{
\item{solution}{stockflow_solution object}

\item{...}{Additional arguments}
}
\value{
Combined plot (if patchwork available) or list of plots
}
\description{
Creates a multi-panel plot showing time series and phase plots.
}
