% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algebraic-api.R
\name{sumvar}
\alias{sumvar}
\title{Create a sum variable specification}
\usage{
sumvar(name, stocks)
}
\arguments{
\item{name}{Character name for the sum variable}

\item{stocks}{Character vector of stock names to sum}
}
\value{
A list with class "sumvar_spec"
}
\description{
Creates a sum variable specification for use with the algebraic API.
}
\examples{
\dontrun{
N <- sumvar("N", stocks = c("S", "E", "I", "R"))
}

}
