% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{SumVariable-class}
\alias{SumVariable-class}
\title{Sum Variable (aggregates stocks)}
\description{
Represents a sum over multiple stocks. Special case of auxiliary variable
commonly used for computing totals (e.g., total population).
}
\section{Slots}{

\describe{
\item{\code{id}}{Integer identifier}

\item{\code{name}}{Character name}

\item{\code{stock_ids}}{Integer vector of stock IDs to sum}
}}

\examples{
\dontrun{
total_pop <- new("SumVariable",
  id = 1L,
  name = "N",
  stock_ids = c(1L, 2L, 3L, 4L)  # S, E, I, R
)
}

}
