% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{to_graphviz}
\alias{to_graphviz}
\title{Export diagram to Graphviz DOT format}
\usage{
to_graphviz(diagram, file = NULL, rankdir = "LR")
}
\arguments{
\item{diagram}{StockFlowDiagram object}

\item{file}{Character file path to write (optional)}

\item{rankdir}{Character direction: "LR" (left-right) or "TB" (top-bottom)}
}
\value{
Character string with DOT code
}
\description{
Creates a DOT language representation that can be rendered with Graphviz.
}
\examples{
\dontrun{
sir <- stock_flow_diagram() \%>\%
  add_stock("S", initial = 999) \%>\%
  add_stock("I", initial = 1) \%>\%
  add_flow("infection", from = "S", to = "I",
    rate = function(inputs, params) params$beta * inputs$S * inputs$I)

# Get DOT string
dot_str <- to_graphviz(sir)
cat(dot_str)

# Write to file
to_graphviz(sir, file = "sir.dot")
# Then: dot -Tpng sir.dot -o sir.png
}

}
