% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{validate_functor}
\alias{validate_functor}
\title{Check if morphism maps are valid}
\usage{
validate_functor(diagram)
}
\arguments{
\item{diagram}{StockFlowDiagram object}
}
\value{
TRUE if valid, character string with error message otherwise
}
\description{
Check if morphism maps are valid
}
\keyword{internal}
