% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algebraic-api.R
\name{variable}
\alias{variable}
\title{Create a variable specification}
\usage{
variable(name, expression)
}
\arguments{
\item{name}{Character name for the variable}

\item{expression}{Function computing variable value}
}
\value{
A list with class "variable_spec"
}
\description{
Creates a variable specification for use with the algebraic API.
}
\examples{
\dontrun{
N <- sumvar("N", stocks = c("S", "E", "I", "R"))
}

}
